+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 
 ## [Unreleased]
 ### Added
+- 2021-06-21 : Bug fixes [liaoxy-idea-helper-v0.0.2](https://github.com/liaoxiangyun/liaoxy-idea-helper/releases/tag/0.0.2)
 - 2021-06-17 : Git Commit Template [liaoxy-idea-helper-v0.0.1](https://github.com/liaoxiangyun/liaoxy-idea-helper/releases/tag/0.0.1)